# openapi/adventure-book-api.yml
openapi: 3.0.3
info:
  title: Adventure Book API
  version: 1.0.0
  description: REST API for browsing and managing adventure books
  contact:
    name: Project Team
    email: chemsklt@example.com

externalDocs:
  description: Adventure Book Wiki
  url: https://wiki/adventure-book


servers:
  - url: http://localhost:8080

tags:
  - name: Books
    description: Book catalog operations
  - name: Book Reading
    description: Book reading operations
  - name: Games
    description: Game session operations

paths:
  /books:
    get:
      tags:
        - Books
      summary: List books with optional filters
      description: Returns all books or filters them by title, author, category, or difficulty.
      operationId: listBooks
      parameters:
        - name: title
          in: query
          required: false
          schema:
            type: string
        - name: author
          in: query
          required: false
          schema:
            type: string
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: difficulty
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Difficulty'
      responses:
        '200':
          description: Books retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookSummaryListResponse'
    post:
      tags:
        - Books
      summary: Add a new book to the collection
      description: Creates a new adventure book with its full structure.
      operationId: createBook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookRequest'
      responses:
        '201':
          description: Book created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookDetailsResponse'
        '400':
          description: Invalid book payload or invalid book structure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Book with the same id already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /books/{bookId}:
    get:
      tags:
        - Books
      summary: Get book details
      description: Returns book details by id.
      operationId: getBookById
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Book details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookDetailsResponse'
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /books/{bookId}/categories:
    post:
      tags:
        - Books
      summary: Add a category to a book
      description: Adds a category to the specified book.
      operationId: addCategoryToBook
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
      responses:
        '200':
          description: Category added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookDetailsResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /books/{bookId}/categories/{categoryName}:
    delete:
      tags:
        - Books
      summary: Remove a category from a book
      description: Removes a category from the specified book.
      operationId: removeCategoryFromBook
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
        - name: categoryName
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Category removed successfully
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /books/{bookId}/start:
    get:
      tags:
        - Book Reading
      summary: Start reading a book
      operationId: startBook
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: First section
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SectionResponse'

  /books/{bookId}/sections/{sectionId}:
    get:
      tags:
        - Book Reading
      summary: Get a section by id
      operationId: getSection
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
        - name: sectionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Section details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SectionResponse'

  /books/{bookId}/sections/{sectionId}/choose:
    post:
      tags:
        - Book Reading
      summary: Choose an option and move to next section
      operationId: chooseOption
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
        - name: sectionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChooseOptionRequest'
      responses:
        '200':
          description: Next section
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SectionResponse'
  /games:
    post:
      tags:
        - Games
      summary: Start a new game session
      operationId: startGame
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartGameRequest'
      responses:
        '201':
          description: Game session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameResponse'
        '400':
          description: Invalid book
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /games/{gameId}:
    get:
      tags:
        - Games
      summary: Get current game session state
      operationId: getGame
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Game session state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameResponse'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /games/{gameId}/choices:
    post:
      tags:
        - Games
      summary: Choose an option in the current game session
      operationId: chooseGameOption
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameChoiceRequest'
      responses:
        '200':
          description: Game state updated after applying the selected option
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameResponse'
        '400':
          description: Invalid game action
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /games/{gameId}/pause:
    post:
      tags:
        - Games
      summary: Pause a game session
      operationId: pauseGame
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Game session paused
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameResponse'
        '400':
          description: Invalid game state transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /games/{gameId}/resume:
    post:
      tags:
        - Games
      summary: Resume a paused game session
      operationId: resumeGame
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Game session resumed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameResponse'
        '400':
          description: Invalid game state transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    Difficulty:
      type: string
      enum:
        - EASY
        - MEDIUM
        - HARD

    BookSummaryResponse:
      type: object
      required:
        - id
        - title
        - author
        - difficulty
        - categories
      properties:
        id:
          type: string
          example: dragon-quest
        title:
          type: string
          example: Dragon Quest
        author:
          type: string
          example: John Doe
        difficulty:
          $ref: '#/components/schemas/Difficulty'
        categories:
          type: array
          items:
            type: string

    BookSummaryListResponse:
      type: object
      required:
        - content
        - totalElements
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/BookSummaryResponse'
        totalElements:
          type: integer
          format: int64
          example: 4

    BookDetailsResponse:
      type: object
      required:
        - id
        - title
        - author
        - difficulty
        - categories
        - sectionsCount
      properties:
        id:
          type: string
        title:
          type: string
        author:
          type: string
        difficulty:
          $ref: '#/components/schemas/Difficulty'
        categories:
          type: array
          items:
            type: string
        sectionsCount:
          type: integer
          example: 12
    CategoryRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          example: Fantasy
    SectionResponse:
      type: object
      required:
        - sectionId
        - text
        - type
      properties:
        sectionId:
          type: string
          example: "1"
        text:
          type: string
        type:
          type: string
          enum: [ BEGIN, NODE, END ]
        options:
          type: array
          items:
            $ref: '#/components/schemas/OptionResponse'
    OptionResponse:
      type: object
      required:
        - id
        - text
        - nextSectionId
      properties:
        id:
          type: string
        text:
          type: string
        nextSectionId:
          type: string
    ChooseOptionRequest:
      type: object
      required:
        - optionId
      properties:
        optionId:
          type: string
    GameStatus:
      type: string
      enum:
        - IN_PROGRESS
        - PAUSED
        - DEAD
        - FINISHED
    ConsequenceType:
      type: string
      enum:
        - LOSE_HEALTH
        - GAIN_HEALTH
    ConsequenceResponse:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/ConsequenceType'
        value:
          type: integer
        text:
          type: string
    StartGameRequest:
      type: object
      required:
        - bookId
      properties:
        bookId:
          type: string
          example: the-prisoner
    GameChoiceRequest:
      type: object
      required:
        - optionId
      properties:
        optionId:
          type: string
          example: "0"
    GameResponse:
      type: object
      required:
        - gameId
        - bookId
        - health
        - status
        - currentSection
      properties:
        gameId:
          type: string
          example: game-123
        bookId:
          type: string
          example: the-prisoner
        health:
          type: integer
          example: 10
        status:
          $ref: '#/components/schemas/GameStatus'
        appliedConsequence:
          $ref: '#/components/schemas/ConsequenceResponse'
        currentSection:
          $ref: '#/components/schemas/SectionResponse'

    CreateConsequenceRequest:
      type: object
      required:
        - type
        - value
        - text
      properties:
        type:
          $ref: '#/components/schemas/ConsequenceType'
        value:
          type: integer
          example: 3
        text:
          type: string
          example: You cut yourself on a rusty nail.

    CreateOptionRequest:
      type: object
      required:
        - description
        - gotoId
      properties:
        description:
          type: string
          example: Open the door
        gotoId:
          type: string
          example: "1000"
        consequence:
          $ref: '#/components/schemas/CreateConsequenceRequest'

    CreateSectionRequest:
      type: object
      required:
        - id
        - text
        - type
      properties:
        id:
          type: string
          example: "1"
        text:
          type: string
          example: You wake up in a dark prison cell.
        type:
          type: string
          enum:
            - BEGIN
            - NODE
            - END
        options:
          type: array
          items:
            $ref: '#/components/schemas/CreateOptionRequest'

    CreateBookRequest:
      type: object
      required:
        - id
        - title
        - author
        - difficulty
        - categories
        - sections
      properties:
        id:
          type: string
          example: the-prisoner
        title:
          type: string
          example: The Prisoner
        author:
          type: string
          example: Daniel El Fuego
        difficulty:
          $ref: '#/components/schemas/Difficulty'
        categories:
          type: array
          items:
            type: string
          example:
            - Horror
            - Mystery
        sections:
          type: array
          items:
            $ref: '#/components/schemas/CreateSectionRequest'

    ErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 404
        error:
          type: string
          example: Not Found
        message:
          type: string
          example: Book with id 'unknown-book' was not found
        path:
          type: string
          example: /books/unknown-book